[FD][APB-1014,APB-1288] Quick fix for behaviour change introduced in APB-1014
Prior to APB-1014 if the user chose a non-UK country we ignored their choice and sent "GB" to the DES subscription API anyway.
APB-1014 changed that so now we send the country sent by the user.
This changes the behaviour back so we always send "GB" again.

This is an interim fix - when we do APB-1288 we willnot silently ignore the country chosen by the user.